+++ spec/database_adapters/mongoid/database_cleaner.rb
@@ -1,9 +1,9 @@
-require 'database_cleaner'
+require 'database_cleaner/mongoid'
 
 RSpec.configure do |config|
   config.before :suite do
-    DatabaseCleaner.strategy = :truncation
-    DatabaseCleaner.clean_with :truncation
+    DatabaseCleaner.strategy = :deletion
+    DatabaseCleaner.clean_with :deletion
   end
 
   config.around :each do |example|