+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 2.1.2 (Next)
 
 * Your contribution here.
+* [#175](https://github.com/slack-ruby/slack-ruby-bot-server/pull/175): Fix(activerecord): correctly check for database in rails 7.2+ - [@markokajzer](https://github.com/markokajzer).
 
 ### 2.1.1 (2023/07/25)